renamed the Home component to Wallet
