+++ examples/wallet/app/constants/routes.json
@@ -1,6 +1,6 @@
 {
   "INDEX": "/",
-  "HOME": "/home",
+  "WALLET": "/wallet",
   "CHOOSE_RESTORE_OR_IMPORT": "/choose_restore_or_import",
   "INPUT_KEYS": "/input_keys",
   "SEND": "/send",
