+++ examples/wallet/app/pages/Wallet.js
